+++ chain/client/src/client_actor.rs
@@ -1957,6 +1957,7 @@ impl ClientActorInner {
             &self.client.chain,
             highest_height,
             &self.network_info.highest_height_peers,
+            self.client.config.sync_max_block_requests,
         )?;
         Ok(block_sync_result)
     }